+++ RELEASE-NOTES.md
@@ -20,6 +20,8 @@
 ## Deprecations
 
 ## New additions
+* Added connection option `--token-file-path` allowing passing OAuth token using a file. The function is also
+  supported by setting `token_file_path` in connection definition.
 
 ## Fixes and improvements
 * The `snow app run` command now allows upgrading to unversioned mode from a versioned or release mode application installation